move mytmpdir() to ananse.utils.py
